Reset first frame if motion detected more than X iterations
Given that we're going to be taking action based on the motion boolean, the
goal here is to reset the base frame if motion persists for too long. This
protects whatever actions we plan to take (screenshots, noises, etc.) from
being played indefinitely if something in the frame permanently changes (e.g.
if the lighting in the room changes over the course of the day).